+++ src/main/java/org/musicplace/global/security/CustomAuthenticationToken.java
@@ -1,4 +1,4 @@
-package org.musicplace.jwt;
+package org.musicplace.global.security;
 
 import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
