+++ src/main/java/org/musicplace/global/security/CustomLoginAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package org.musicplace.jwt;
+package org.musicplace.global.security;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.AuthenticationException;
